Split song actions into own component. Added song deletion with confirmation modal
