+++ src/components/app/Pages/SongEditPage.svelte
@@ -5,6 +5,7 @@
   import { User, Doc } from "sveltefire";
   import AppHeader from "../../ui/AppHeader.svelte";
   import SongEdit from "../SongEdit/SongEdit.svelte";
+  import SongEditActions from "../SongEdit/SongEditActions.svelte";
   import VisuallyHidden from "../../ui/VisuallyHidden.svelte";
   export let id;
   import { ButtonText, ButtonLink } from "../../ui/Button";
@@ -19,7 +20,7 @@
 
 <Modal>
   <User persist={sessionStorage} let:user>
-    <Doc path={`/users/${user.uid}/songs/${id}`} let:data={song} let:ref={songRef}>
+    <Doc path={`/users/${user.uid}/songs/${id}`} let:data={song} let:ref={docRef}>
       <AppHeader title="Edit song">
         <div slot="start" in:fade>
           <ButtonLink to="/songs/{id}">
@@ -28,13 +29,14 @@
           </ButtonLink>
         </div>
         <div slot="end" in:fade>
-          <ButtonText on:click={() => handleSave(song, songRef)}
+          <ButtonText on:click={() => handleSave(song, docRef)}
             >Save</ButtonText
           >
         </div>
       </AppHeader>
       <div class="page" in:fade>
         <SongEdit {song} isEdit="true" />
+        <SongEditActions {docRef} />
       </div>
       <div slot="loading">Loading...</div>
     </Doc>
