+++ src/components/app/SongEdit/SongEdit.svelte
@@ -8,33 +8,13 @@
   } from "../../ui/Form";
   import SongEditDetails from "./SongEditDetails.svelte";
   import SongEditChords from "./SongEditChords.svelte";
-  import ButtonDefault from "../../ui/Button/ButtonDefault.svelte";
-  import Heading from "../../ui/Heading/Heading.svelte";
   
   export let song = {};
-  export let isEdit = false;
-
-  function handleDelete() {
-    // TODO
-  }
 </script>
 
 <style>
   .wrapper {
     padding: 1rem;
-  }
-  .footer {
-    display: flex;
-    align-items: center;
-    padding: 1rem;
-    background-color: var(--neutralLight);
-    box-shadow: inset 0 1px 8px rgba(0, 0, 0, 0.25)
-  }
-
-  .footer :global(button) {
-    --buttonColor: var(--danger);
-    --buttonBorderColor: var(--danger);
-    margin-left: 1rem;
   }
 </style>
 
@@ -70,10 +50,4 @@
       tuning={song.tuning}
       capoAdjustment={song.capoAdjustment} />
   </div>
-  {#if isEdit}
-  <div class="footer">
-    <Heading text="Actions:" level="2" fontSize="1rem"/>
-    <ButtonDefault on:click={handleDelete}>Delete song</ButtonDefault>
-  </div>
-  {/if}
 </Fieldset>
